% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{- is a data frame that should contain two variables in space i(0,1,2...N): w - weight, and v - value. Theese should be positive values only.}

\item{W}{- is the maximum weight for the knapsack. Should be a positive integer.}
}
\value{
maximum value for weight
}
\description{
Greedy
}
\examples{
set.seed(42)
n <- 2000
knapsack_objects <- data.frame(w = sample(1:4000, size = n, replace = TRUE),
                              v <- runif(n = n, 0, 10000))
}
