% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack.R
\docType{data}
\name{knapsack_objects}
\alias{knapsack_objects}
\title{knapsack_objects}
\format{
An object of class \code{data.frame} with 2000 rows and 2 columns.
}
\usage{
knapsack_objects
}
\value{
- an example data frame for the knapsack.problem package
}
\description{
- Loads an example data frame to use with the functions of the knapsack.problem package. w is made by sample(1:400, 2000, T) with seed(42) and v is a sample from the uniform distribution of the same size.
}
\keyword{datasets}
